NamDD.SP-984/feat/add-title-and-desc fix title and description
