+++ src/app/assets/styles/intro.scss
@@ -28,6 +28,7 @@
   }
 
   > img {
+    margin-bottom: 50px;
     margin-top: 90px;
     position: relative;
   }
